+++ src/main/java/com/example/demo/repository/CarritoProductoRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.entity.CarritoProducto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CarritoProductoRepository extends JpaRepository<CarritoProducto, Integer> {
+public interface CarritoProductoRepository extends JpaRepository<CarritoProducto, Long> {
 }
