+++ src/main/java/com/example/demo/repository/CarritoRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.entity.Carrito;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CarritoRepository extends JpaRepository <Carrito, Integer>{
+public interface CarritoRepository extends JpaRepository <Carrito, Long>{
 }
