+++ src/main/java/com/example/demo/repository/CompraProductoRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.entity.CompraProducto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CompraProductoRepository extends JpaRepository<CompraProducto, Integer> {
+public interface CompraProductoRepository extends JpaRepository<CompraProducto, Long> {
 }
