+++ src/main/java/com/example/demo/repository/ListaDeDeseosRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.entity.ListaDeDeseos;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ListaDeDeseosRepository extends JpaRepository<ListaDeDeseos, Integer>{
+public interface ListaDeDeseosRepository extends JpaRepository<ListaDeDeseos, Long>{
 }
