+++ src/main/java/com/example/demo/repository/ListaProductoRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.entity.ListaProducto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ListaProductoRepository extends JpaRepository<ListaProducto, Integer> {
+public interface ListaProductoRepository extends JpaRepository<ListaProducto, Long> {
 }
