+++ src/main/java/com/example/demo/repository/TiendaRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.entity.Tienda;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TiendaRepository extends JpaRepository <Tienda, Integer>{
+public interface TiendaRepository extends JpaRepository <Tienda, Long>{
 }
