+++ src/main/java/com/example/demo/repository/ValoracionRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.entity.Valoracion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ValoracionRepository extends JpaRepository<Valoracion, Integer> {
+public interface ValoracionRepository extends JpaRepository<Valoracion, Long> {
 }
